fix gem dependencies
Downloading loofah-2.3.1 revealed dependencies not in the API or the lockfile (nokogiri (>= 1.5.9)).
Either installing with `--full-index` or running `bundle update loofah` should fix the problem.

Downloading rails-html-sanitizer-1.0.4 revealed dependencies not in the API or the lockfile (loofah (~> 2.2, >=
2.2.2)).
Either installing with `--full-index` or running `bundle update rails-html-sanitizer` should fix the problem.

Downloading railties-6.0.0.rc1 revealed dependencies not in the API or the lockfile (rake (>= 0.8.7)).
Either installing with `--full-index` or running `bundle update railties` should fix the problem.