Add compatability for ; in sqlalchemy,psycopg2 & django (#123)
This PR fixes a zero day issue where comments are added after the end of SQL statement i.e. after `;`.  This will create no or wrong association of the comment wrt the SQL statement.
We previously fixed it for psycopg2 only, but this CL reactors the code and fixes the issue for sqlalchemy and as a precaution was added to django also.