Added dummy workflow tests for non laravel code changes (#135)
Add Dummy tests for all libraries except laravel-sqlcommenter.

When we are adding status check to checkin code for PHP SQLCommenter, the same status check is added for all other libraries also, as this is a mono-repo. 
To avoid that, we are following a strategy of selective filtering instead and tests will have same name but selectively filter what to run on the basis of path.