fix: handle Starlette routes in FastAPI middleware (#136)
This commit changes the FastAPI middleware so that the `"route"` info is
now optional. If it is not present, the SQL comment will not have the
controller/route info.

This info is always present in FastAPI routes. However, Starlette-based
routes (FastAPI can mount Starlette apps as sub-applications) do not and
would crash when the middleware tried to extract the information.

Fixes #133